requirements.txt
selenium
allure-behave
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
google-auth